requirements.txt
Django==1.8.4
django-braces==1.8.1
django-crispy-forms==1.5.2
